ft: implement bussiness logic for mutation resolvers #2

+++ models/models.go
@@ -5,9 +5,9 @@ type Session struct {
 	Owner       string `json:"owner"`
 	Title       string `json:"title"`
 	Description string `json:"description"`
-	Start       int    `json:"start"`
-	End         int    `json:"end"`
-	Ts          int    `json:"Ts"`
+	Start       int64    `json:"start"`
+	End         int64    `json:"end"`
+	Ts          int64    `json:"Ts"`
 }
 
 type User struct {
@@ -15,5 +15,5 @@ type User struct {
 	Name     string `json:"name"`
 	Email    string `json:"email"`
 	Password string `json:"password"`
-	Ts       int    `json:"Ts"`
+	Ts       int64    `json:"Ts"`
 }
